Remove palette TUI. Palette editing only available in the project viewer/UI.
