--- tests/tui/test_edit_features.py
@@ -1,212 +1,0 @@
-#!/usr/bin/env python3
-"""
-Test script for the new color editing features in the Palette TUI.
-
-This script tests:
-1. Live color preview in the edit dialog
-2. Color validation
-3. Edit dialog composition
-"""
-
-from pathlib import Path
-
-from rich.style import Style
-from rich.text import Text
-
-
-def test_color_edit_screen() -> None:
-    """Test the color validation function."""
-    from stride.ui.tui import validate_color
-
-    print("Testing color validation...")
-
-    # Test validation
-    test_cases = [
-        ("#FF5733", True),
-        ("#FF5733FF", True),
-        ("rgb(255, 87, 51)", True),
-        ("rgba(255, 87, 51, 1.0)", True),
-        ("rgb(255,87,51)", True),  # No spaces
-        ("#1E90FF", True),
-        ("invalid", False),
-        ("#GGG", False),
-        ("blue", False),
-        ("", False),
-    ]
-
-    print("\n  Color validation tests:")
-    for color, expected_valid in test_cases:
-        result = validate_color(color)
-        status = "✓" if result == expected_valid else "✗"
-        print(f"    {status} '{color}': {result}")
-
-    print("\n  ✓ Color validation tests passed!")
-
-
-def test_live_preview_simulation() -> None:
-    """Simulate the live preview behavior."""
-    from stride.ui.tui import color_to_rich_format
-
-    print("\nTesting live preview simulation...")
-
-    colors = ["#FF5733", "#1E90FF", "rgb(26, 188, 156)", "rgba(255, 87, 51, 0.5)"]
-
-    print("  Simulating color changes:")
-    for color in colors:
-        # Simulate what happens in on_input_changed
-        rich_color = color_to_rich_format(color)
-        preview = Text("████████████", style=Style(color=rich_color))
-        print(f"    Preview for '{color}' -> '{rich_color}': {preview}")
-
-    print("✓ Live preview simulation test passed!")
-
-
-def test_cursor_styling() -> None:
-    """Test that cursor styling is configured."""
-    from stride.ui.tui import PaletteViewer
-
-    print("\nTesting cursor styling configuration...")
-
-    # Check that CSS includes cursor styling
-    css = PaletteViewer.CSS
-    if "cursor-background" in css or "cursor" in css:
-        print("  ✓ Cursor styling found in CSS")
-    else:
-        print("  ⚠ No cursor styling found in CSS")
-
-    print("  ✓ Cursor styling test complete!")
-
-
-def test_full_edit_workflow() -> None:
-    """Test the complete edit workflow."""
-    from stride.models import ProjectConfig
-    from stride.ui.tui import organize_palette_by_groups
-
-    print("\nTesting full edit workflow...")
-
-    # Load test project
-    project_path = Path("test_project/project.json5")
-    if not project_path.exists():
-        print("  ⚠ Test project not found, skipping workflow test")
-        return
-
-    config = ProjectConfig.from_file(project_path)
-    print(f"  ✓ Loaded project: {config.project_id}")
-
-    # Organize palette
-    groups = organize_palette_by_groups(config.color_palette, config)
-    print(f"  ✓ Organized into {len(groups)} groups")
-
-    # Simulate editing a color
-    test_label = "residential"
-    if any(test_label in labels for labels in groups.values()):
-        print(f"  ✓ Found '{test_label}' in palette")
-
-        # Find the current color
-        for group_name, labels in groups.items():
-            if test_label in labels:
-                old_color = labels[test_label]
-                print(f"    Current color: {old_color}")
-
-                # Simulate editing
-                new_color = "#1E90FF"
-                labels[test_label] = new_color
-                print(f"    New color: {new_color}")
-
-                # Verify the change
-                assert labels[test_label] == new_color
-                print("    ✓ Color updated successfully")
-                break
-    else:
-        print(f"  ⚠ '{test_label}' not found in palette")
-
-    print("  ✓ Full edit workflow test passed!")
-
-
-def test_color_preview_widget() -> None:
-    """Test the color preview widget rendering."""
-    from stride.ui.tui import color_to_rich_format
-
-    print("\nTesting color preview widget...")
-
-    colors = [
-        "#FF5733",
-        "#1E90FF",
-        "rgb(26, 188, 156)",
-        "rgba(255, 87, 51, 0.8)",
-    ]
-
-    print("  Creating preview widgets:")
-    for color in colors:
-        rich_color = color_to_rich_format(color)
-        print(f"    ✓ Created preview for {color} -> {rich_color}")
-
-    print("✓ Color preview widget test passed!")
-
-
-def test_edit_dialog_composition() -> None:
-    """Test that the PaletteViewer can be instantiated."""
-    from pathlib import Path
-
-    from stride.ui.tui import PaletteViewer
-
-    print("\nTesting PaletteViewer instantiation...")
-
-    # Create a simple test palette with label groups
-    test_label_groups = {"Test Group": {"test_label": "#FF5733", "another_label": "#1E90FF"}}
-
-    viewer = PaletteViewer(
-        palette_name="test_palette",
-        palette_location=Path("/tmp/test_palette.json"),
-        palette_type="user",
-        label_groups=test_label_groups,
-    )
-    print("  ✓ Created PaletteViewer instance")
-
-    # Verify basic attributes
-    assert viewer.palette_name == "test_palette"
-    assert viewer.palette_type == "user"
-    print("  ✓ Palette name and type set correctly")
-
-    # Note: Can't call compose() outside of app context
-    print("  ℹ Skipping compose() test (requires app context)")
-
-    print("  ✓ PaletteViewer instantiation test passed!")
-
-
-def main() -> int:
-    """Run all tests."""
-    print("=" * 60)
-    print("Testing Palette TUI Edit Features")
-    print("=" * 60)
-
-    try:
-        test_color_edit_screen()
-        test_live_preview_simulation()
-        test_cursor_styling()
-        test_color_preview_widget()
-        test_edit_dialog_composition()
-        test_full_edit_workflow()
-
-        print("\n" + "=" * 60)
-        print("✓ All edit feature tests passed!")
-        print("=" * 60)
-        print("\nTo test interactively:")
-        print("  stride palette view test_project --project")
-        print("  - Navigate to a color with arrow keys")
-        print("  - Press 'e' to see the live preview")
-        print("  - Type different color values to see preview update")
-        print("  - The cursor should be lighter and not mask colors")
-
-    except Exception as e:
-        print(f"\n✗ Test failed with error: {e}")
-        import traceback
-
-        traceback.print_exc()
-        return 1
-
-    return 0
-
-
-if __name__ == "__main__":
-    exit(main())
